requirements.txt
opencv-python==4.11.0.86
mediapipe==0.10.21
pyserial==3.5